requirements.txt
alpaca-py>=0.29
python-dotenv>=1.0
pandas>=2.0
